src/week5/requirements.txt
openai
streamlit
langchain
python-docx 
langchain_experimental
pyodbc
matplotlib

